requirements.txt
numpy==1.26.3
onnxruntime==1.16.3
gradio==3.50.2
Pillow==10.1.0
huggingface-hub==0.20.1
pydantic==1.10.13